pyproject.toml
[project]
name = "simple-hr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fpdf>=1.7.2",
    "google-genai>=1.35.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.6.7",
    "markdown2>=2.5.4",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "tavily-python>=0.7.12",
    "xhtml2pdf>=0.2.17",
]
